Finish Backend development

+++ backend/src/database/migrations/20221114234511-account.js
@@ -10,7 +10,7 @@ module.exports = {
       
       balance: {
         allowNull: false,
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(10,2)
       }
     
     });
